requirements.txt
Flask
Flask-SQLAlchemy
bcrypt
redis
celery[redis]
flask-sse
gunicorn
gevent
Flask-Caching
email_validator
python-dotenv
Flask-Cors
flask_restful
Authlib 
requests
flask-bcrypt 
PyJWT
pillow
pytest
flask-sse
flask-mail
